'*' read permissions means chmod o+rx
If a repository gives permission for everyone, then it is replicated to the filesystem giving read access to everyone through chmod. This way, if there is a Webgit running on other user, with will list only the repositories that everyone can already read.
Note that on linux, when you create a dir is permissions are 'o+rx' by default. So this new behavior is more restrictive than the previous behavior.